+++ mod.ts
@@ -1,4 +1,4 @@
 export type { SendConfig } from "./config/mail/mod.ts";
-export type { ClientOptions } from "./config/client/mod.ts";
+export type { ClientOptions } from "./config/client.ts";
 export { SMTPHandler as SMTPClient } from "./client/mod.ts";
 export { quotedPrintableEncode } from "./config/mail/encoding.ts";
